#419 - Removed rounding to integers - use double instead

+++ dkpro-tc-ml-liblinear/src/main/java/org/dkpro/tc/ml/liblinear/LiblinearTestTask.java
@@ -138,7 +138,7 @@ public class LiblinearTestTask extends ExecutableTaskBase implements Constants {
 			Feature[] instance = testInstances[i];
 			Double prediction = Linear.predict(model, instance);
 
-			writer.write(prediction.intValue() + SEPARATOR_CHAR + new Double(test.y[i]).intValue());
+			writer.write(prediction + SEPARATOR_CHAR + new Double(test.y[i]));
 			writer.write("\n");
 		}
